Run taxon search when pressing enter from text search input
